+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoView.java
@@ -61,6 +61,7 @@ import java.io.InputStream;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.ref.WeakReference;
+import java.util.HashMap;
 import java.util.Objects;
 import org.mozilla.gecko.AndroidGamepadManager;
 import org.mozilla.gecko.EventDispatcher;
@@ -72,6 +73,10 @@ import org.mozilla.gecko.util.ThreadUtils;
 public class GeckoView extends FrameLayout implements GeckoDisplay.NewSurfaceProvider {
   private static final String LOGTAG = "GeckoView";
   private static final boolean DEBUG = false;
+  private static final String KEYBOARD_WINDOW_INSETS_LISTENER = "KEYBOARD_WINDOW_INSETS_LISTENER";
+  private final HashMap<String, androidx.core.view.OnApplyWindowInsetsListener>
+      mWindowInsetsListeners =
+          new HashMap<String, androidx.core.view.OnApplyWindowInsetsListener>();
 
   protected final @NonNull Display mDisplay = new Display();
 
@@ -199,34 +204,10 @@ public class GeckoView extends FrameLayout implements GeckoDisplay.NewSurfacePro
     @Override // OnApplyWindowInsetsListener
     public WindowInsetsCompat onApplyWindowInsets(
         @NonNull final View view, @NonNull final WindowInsetsCompat insets) {
-      final WindowInsetsCompat updatedInsets =
-          WindowInsetsCompat.toWindowInsetsCompat(
-              view.onApplyWindowInsets(insets.toWindowInsets()));
       if (mDisplay != null) {
-        mDisplay.windowInsetsChanged(updatedInsets);
+        mDisplay.windowInsetsChanged(insets);
       }
-      return updatedInsets;
-    }
-
-    // Attach this class as a WindowInsets listener to tell the software keyboard
-    // changes.
-    public void attachWindowInsetsListener(final Activity activity) {
-      try {
-        final View rootView = activity.getWindow().getDecorView().getRootView();
-        ViewCompat.setOnApplyWindowInsetsListener(rootView, this);
-      } catch (final Exception e) {
-        Log.e(LOGTAG, "Failed to attach WindowInsetsListener: ", e);
-      }
-    }
-
-    // Detach as the listener.
-    public void detachWindowInsetsListener(final Activity activity) {
-      try {
-        final View rootView = activity.getWindow().getDecorView().getRootView();
-        ViewCompat.setOnApplyWindowInsetsListener(rootView, null);
-      } catch (final Exception e) {
-        Log.e(LOGTAG, "Failed to detach WindowInsetsListener: ", e);
-      }
+      return insets;
     }
 
     public boolean shouldPinOnScreen() {
@@ -647,6 +628,74 @@ public class GeckoView extends FrameLayout implements GeckoDisplay.NewSurfacePro
     return mSession.getPanZoomController();
   }
 
+  /**
+   * Register an internal windowInsetsListener that will forward its calls to the listeners
+   * registered in {@link GeckoView#mWindowInsetsListeners}
+   *
+   * @param activity The target Activity to observe.
+   */
+  private void attachWindowInsetsListener(final @NonNull Activity activity) {
+    try {
+      final View rootView = activity.getWindow().getDecorView().getRootView();
+      ViewCompat.setOnApplyWindowInsetsListener(
+          rootView,
+          (view, insets) -> {
+            WindowInsetsCompat updatedInsets =
+                WindowInsetsCompat.toWindowInsetsCompat(
+                    view.onApplyWindowInsets(insets.toWindowInsets()));
+
+            for (final androidx.core.view.OnApplyWindowInsetsListener listener :
+                mWindowInsetsListeners.values()) {
+              updatedInsets = listener.onApplyWindowInsets(view, updatedInsets);
+            }
+            return updatedInsets;
+          });
+    } catch (final Exception e) {
+      Log.e(LOGTAG, "Failed to attach WindowInsetsListener: ", e);
+    }
+  }
+
+  /**
+   * Unregister the internal WindowInsetsListener attached to the Activity's root view and clear the
+   * listeners that were attached through {@link GeckoView#addWindowInsetsListener(String,
+   * androidx.core.view.OnApplyWindowInsetsListener) }
+   *
+   * @param activity The target Activity to stop observing.
+   */
+  private void detachWindowInsetsListener(final @NonNull Activity activity) {
+    try {
+      final View rootView = activity.getWindow().getDecorView().getRootView();
+      ViewCompat.setOnApplyWindowInsetsListener(rootView, null);
+    } catch (final Exception e) {
+      Log.e(LOGTAG, "Failed to detach WindowInsetsListener: ", e);
+    }
+  }
+
+  /**
+   * Add an OnApplyWindowInsetsListener to observe the root view WindowInsets changes.
+   *
+   * @param key The key associated to the listener.
+   * @param listener The OnApplyWindowInsetsListener to be invoked.
+   */
+  @UiThread
+  public void addWindowInsetsListener(
+      final @NonNull String key,
+      final @Nullable androidx.core.view.OnApplyWindowInsetsListener listener) {
+    ThreadUtils.assertOnUiThread();
+    mWindowInsetsListeners.put(key, listener);
+  }
+
+  /**
+   * Remove the OnApplyWindowInsetsListener to stop observing WindowInsets changed.
+   *
+   * @param key The key associated to the listener to remove.
+   */
+  @UiThread
+  public void removeWindowInsetsListener(final @NonNull String key) {
+    ThreadUtils.assertOnUiThread();
+    mWindowInsetsListeners.remove(key);
+  }
+
   @Override
   public void onAttachedToWindow() {
     if (mIsSessionPoisoned) {
@@ -666,14 +715,18 @@ public class GeckoView extends FrameLayout implements GeckoDisplay.NewSurfacePro
     super.onAttachedToWindow();
 
     // This needs to be called after the `super.onAttachedToWindow()`.
-    mDisplay.attachWindowInsetsListener(getActivityFromContext(getContext()));
+    addWindowInsetsListener(KEYBOARD_WINDOW_INSETS_LISTENER, mDisplay);
+    attachWindowInsetsListener(getActivityFromContext(getContext()));
   }
 
   @Override
   public void onDetachedFromWindow() {
     super.onDetachedFromWindow();
 
-    mDisplay.detachWindowInsetsListener(getActivityFromContext(getContext()));
+    removeWindowInsetsListener(KEYBOARD_WINDOW_INSETS_LISTENER);
+    if (mWindowInsetsListeners.isEmpty()) {
+      detachWindowInsetsListener(getActivityFromContext(getContext()));
+    }
 
     if (mSession == null) {
       return;